test: split it/utils.rs into modules
Moves the 4.5k-line utils.rs into tests/it/utils/ with one module per
concern: node (ZoneTestNode + mock L1 RPC), l1 (L1TestNode, TIP-403
seeding, genesis builders), accounts (ZoneAccount, withdrawal args),
p2p (cluster + sequencer spawning), fixture (L1Fixture, queue
injection), private_rpc (auth tokens, RPC contexts), and a mod.rs
facade that re-exports everything pub(crate) — no call-site changes in
any test file.

Move-only, with three mechanical adjustments: the #[path] include of
auth_tokens.rs gains one directory level, items that now cross module
boundaries are pub(crate) instead of private, and each module keeps
only the imports it uses. Also reattaches ZoneTestNode's doc block,
which had been orphaned onto the RpcApiFuture type alias.

Review with `git diff --color-moved=dimmed-zebra`.
